+++ src/BayesFilters/src/PFPrediction.cpp
@@ -37,13 +37,13 @@ bool PFPrediction::skip(const std::string& what_step, const bool status)
     {
         getStateModel().skip("state", status);
 
-        skip_ = getStateModel().getSkipState() & getStateModel().exogenous_model().getSkipState();
+        skip_ = getStateModel().is_skipping() & getStateModel().exogenous_model().is_skipping();
     }
     else if (what_step == "exogenous")
     {
         getStateModel().skip("exogenous", status);
 
-        skip_ = getStateModel().getSkipState() & getStateModel().exogenous_model().getSkipState();
+        skip_ = getStateModel().is_skipping() & getStateModel().exogenous_model().is_skipping();
     }
     else
         return false;
@@ -52,7 +52,7 @@ bool PFPrediction::skip(const std::string& what_step, const bool status)
 }
 
 
-bool PFPrediction::getSkipState()
+bool PFPrediction::is_skipping()
 {
     return skip_;
 }