+++ src/BayesFilters/src/ParticleFilter.cpp
@@ -25,28 +25,6 @@ noexcept :
 { }
 
 
-ParticleFilter::ParticleFilter(ParticleFilter&& pf) noexcept :
-    initialization_(std::move(pf.initialization_)),
-    prediction_(std::move(pf.prediction_)),
-    correction_(std::move(pf.correction_)),
-    resampling_(std::move(pf.resampling_))
-{ }
-
-
-ParticleFilter& ParticleFilter::operator=(ParticleFilter&& pf) noexcept
-{
-    if (this == &pf)
-        return *this;
-
-    initialization_ = std::move(pf.initialization_);
-    prediction_     = std::move(pf.prediction_);
-    correction_     = std::move(pf.correction_);
-    resampling_     = std::move(pf.resampling_);
-
-    return *this;
-}
-
-
 bool ParticleFilter::skip(const std::string& what_step, const bool status)
 {
     if (what_step == "prediction" ||
@@ -62,8 +40,6 @@ bool ParticleFilter::skip(const std::string& what_step, const bool status)
         bool return_status = true;
 
         return_status &= prediction_->skip("prediction", status);
-        return_status &= prediction_->skip("state", status);
-        return_status &= prediction_->skip("exogenous", status);
 
         return_status &= correction_->skip(status);
 
@@ -71,4 +47,28 @@ bool ParticleFilter::skip(const std::string& what_step, const bool status)
     }
 
     return false;
+}
+
+
+ParticleSetInitialization& ParticleFilter::initialization()
+{
+    return *initialization_;
+}
+
+
+PFPrediction& ParticleFilter::prediction()
+{
+    return *prediction_;
+}
+
+
+PFCorrection& ParticleFilter::correction()
+{
+    return *correction_;
+}
+
+
+Resampling& ParticleFilter::resampling()
+{
+    return *resampling_;
 }